# api/openapi.yml
openapi: 3.0.3
info:
  title: Strava Export API
  description: API for exporting and managing Strava activity data
  version: 0.1.0
  contact:
    name: Danila Loginov
    email: danila@loginov.rocks
    url: https://loginov.rocks
  license:
    name: MIT
    url: https://github.com/loginov-rocks/Strava-Export/blob/main/LICENSE
...
servers:
  - url: http://localhost:3001
    description: Local Server
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /auth/callback:
    get:
      summary: Exchange authorization code for tokens
      description: Exchange OAuth authorization code for access and refresh tokens
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: OAuth authorization code from Strava
                scope:
                  type: array
                  items:
                    type: string
                  description: Optional OAuth scopes
                state:
                  type: string
                  description: Optional state parameter
              required:
                - code
      responses:
        204:
          description: Tokens exchanged successfully
          headers:
            Set-Cookie:
              description: Access and refresh tokens set as HTTP-only cookies
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"

  # TODO: Update
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Authentication via Personal Access Token (for APIs)
    cookieAuth:
      type: apiKey
      in: cookie
      name: stravaExportAccessToken
      description: Authentication via HTTP-only cookies (for Web App)
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
      required:
        - message
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Bad Request
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Unauthorized
...